fix chef13 race for ulimit_domain, and merge custom resources together to fix races with blocks
