+++ metadata.rb
@@ -4,4 +4,4 @@ license          "Apache 2.0"
 description      "Installs/Configures ulimit"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 name             "ulimit"
-version          "0.4.0"
+version          "0.5.0"
